requirements.txt
attrs==19.1.0
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-filter==2.2.0
django-rest-framework==0.1.0
django-rest-registration==0.5.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
drf-yasg==1.16.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
packaging==19.1
psycopg2==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pyparsing==2.4.2
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
